// src/data/redux/worklogs/types.ts
export interface IWorklog {
  _id: string;
  worklogThumbs: {
    _id: string;
  };
  worklogsAreLoading: boolean;
  filters: { [index: string]: number };
}

export const enum IWorklogActions {
  SET_WORKLOGS_BY_BUDGET_ID = "SET_WORKLOGS_BY_BUDGET_ID",
  SET_PREVIOUS_WORKLOGS = "SET_PREVIOUS_WORKLOGS",
  SET_WORKLOGS_BY_BUDGET_ID_FROM_CACHE = "SET_WORKLOGS_BY_BUDGET_ID_FROM_CACHE",
  SET_WORKLOGS_FILTER = "SET_WORKLOGS_FILTER",
  UPDATE_WORKLOGS_BY_BUDGET_ID = "UPDATE_WORKLOGS_BY_BUDGET_ID",
  SET_WORKLOGS_LOADING_STATUS = "SET_WORKLOGS_LOADING_STATUS",
  LOAD_BUDGETS_WORKLOGS = "LOAD_BUDGETS_WORKLOGS",
  LOAD_PREVIOUS_WORKLOGS = "LOAD_PREVIOUS_WORKLOGS",
  SET_BUDGETS_WORKLOGS_TO_REDUX = "SET_BUDGETS_WORKLOGS_TO_REDUX",
  CLEAR_WORKLOGS = "CLEAR_WORKLOGS",
  CLEAR_FILTER = "CLEAR_FILTER"
}